+++ of_to_tp.py
@@ -30,13 +30,16 @@ class PrintTaskpaperVisitor(Visitor):
     def end_folder (self, folder):
         self.depth-=1
     def begin_project (self, project):
-        print >>self.out, self.tabs() + project.name + ':'
+        if self.project_mode:
+            print >>self.out, self.tabs() + project.name + ':'
+        else:
+            print >>self.out, self.tabs() + '- ' + self.remove_trailing_colon(project.name) + self.tags(project)
         self.depth+=1
     def end_project (self, project):
         self.depth-=1
     def begin_task (self, task):
         if self.is_empty (task) or self.project_mode == False:
-            print >>self.out, self.tabs() + '- ' + task.name + self.tags(task)
+            print >>self.out, self.tabs() + '- ' + self.remove_trailing_colon(task.name) + self.tags(task)
         else:
             print >>self.out, self.tabs() + task.name + ':'
         self.depth+=1
@@ -64,6 +67,10 @@ class PrintTaskpaperVisitor(Visitor):
         if len (tags) > 0:
             return ' ' + ' '.join(tags)
         return ''
+    def remove_trailing_colon (self, string):
+        if string.endswith(':'):
+            return string[:-1]
+        return string
     def is_empty (self, item):
         return len ([x for x in item.children if x.marked]) == 0
     def tabs (self):